# Conflicts:
#	pom.xml
#	src/main/java/cn/nukkit/block/BlockGrass.java
#	src/main/java/cn/nukkit/block/BlockMushroom.java
#	src/main/java/cn/nukkit/block/BlockMycelium.java
#	src/main/java/cn/nukkit/block/BlockSapling.java
#	src/main/java/cn/nukkit/block/BlockTNT.java
#	src/main/java/cn/nukkit/item/food/FoodChorusFruit.java
#	src/main/java/cn/nukkit/level/biome/EnumBiome.java
#	src/main/java/cn/nukkit/level/biome/impl/iceplains/IcePlainsSpikesBiome.java
#	src/main/java/cn/nukkit/level/generator/FlatGenerator.java
#	src/main/java/cn/nukkit/level/generator/Generator.java
#	src/main/java/cn/nukkit/level/generator/GeneratorFactory.java
#	src/main/java/cn/nukkit/level/generator/NetherGenerator.java
#	src/main/java/cn/nukkit/level/generator/NormalGenerator.java
#	src/main/java/cn/nukkit/level/generator/SimpleChunkManager.java
#	src/main/java/cn/nukkit/level/generator/function/ChunkGenerateFunction.java
#	src/main/java/cn/nukkit/level/generator/function/ChunkPopulateFunction.java
#	src/main/java/cn/nukkit/level/generator/object/BasicGenerator.java
#	src/main/java/cn/nukkit/level/generator/object/ObjectTallGrass.java
#	src/main/java/cn/nukkit/level/generator/object/mushroom/BigMushroom.java
#	src/main/java/cn/nukkit/level/generator/object/ore/OreType.java
#	src/main/java/cn/nukkit/level/generator/object/tree/HugeTreesGenerator.java
#	src/main/java/cn/nukkit/level/generator/object/tree/NewJungleTree.java
#	src/main/java/cn/nukkit/level/generator/object/tree/ObjectBigSpruceTree.java
#	src/main/java/cn/nukkit/level/generator/object/tree/ObjectDarkOakTree.java
#	src/main/java/cn/nukkit/level/generator/object/tree/ObjectJungleBigTree.java
#	src/main/java/cn/nukkit/level/generator/object/tree/ObjectSavannaTree.java
#	src/main/java/cn/nukkit/level/generator/object/tree/ObjectSpruceTree.java
#	src/main/java/cn/nukkit/level/generator/object/tree/ObjectSwampTree.java
#	src/main/java/cn/nukkit/level/generator/object/tree/ObjectTree.java
#	src/main/java/cn/nukkit/level/generator/object/tree/TreeGenerator.java
#	src/main/java/cn/nukkit/level/generator/populator/impl/MushroomPopulator.java
#	src/main/java/cn/nukkit/level/generator/populator/impl/PopulatorDoublePlant.java
#	src/main/java/cn/nukkit/level/generator/populator/impl/PopulatorGlowStone.java
#	src/main/java/cn/nukkit/level/generator/populator/impl/PopulatorLava.java
#	src/main/java/cn/nukkit/level/generator/populator/impl/PopulatorOre.java
#	src/main/java/cn/nukkit/level/generator/populator/impl/PopulatorSugarcane.java
#	src/main/java/cn/nukkit/level/generator/populator/impl/PopulatorTree.java
#	src/main/java/cn/nukkit/level/generator/populator/impl/tree/DarkOakTreePopulator.java
#	src/main/java/cn/nukkit/level/generator/populator/impl/tree/JungleBigTreePopulator.java
#	src/main/java/cn/nukkit/level/generator/populator/impl/tree/JungleFloorPopulator.java
#	src/main/java/cn/nukkit/level/generator/populator/impl/tree/JungleTreePopulator.java
#	src/main/java/cn/nukkit/level/generator/populator/impl/tree/SavannaTreePopulator.java
#	src/main/java/cn/nukkit/level/generator/populator/impl/tree/SpruceBigTreePopulator.java
#	src/main/java/cn/nukkit/level/generator/populator/impl/tree/SpruceMegaTreePopulator.java
#	src/main/java/cn/nukkit/level/generator/populator/impl/tree/SwampTreePopulator.java
#	src/main/java/cn/nukkit/level/manager/LevelChunkManager.java
#	src/main/java/cn/nukkit/math/BedrockRandom.java
#	src/main/java/cn/nukkit/math/BlockFace.java

+++ src/main/java/cn/nukkit/block/BlockChorusFlower.java
@@ -11,12 +11,12 @@ public class BlockChorusFlower extends BlockTransparent {
     }
 
     @Override
-    public double getHardness() {
-        return 0.4;
+    public float getHardness() {
+        return 0.4f;
     }
 
     @Override
-    public double getResistance() {
+    public float getResistance() {
         return 2;
     }
