+++ src/main/java/cn/nukkit/block/BlockChorusPlant.java
@@ -12,12 +12,12 @@ public class BlockChorusPlant extends BlockTransparent {
     }
 
     @Override
-    public double getHardness() {
-        return 0.4;
+    public float getHardness() {
+        return 0.4f;
     }
 
     @Override
-    public double getResistance() {
+    public float getResistance() {
         return 2;
     }
