+++ src/main/java/cn/nukkit/block/BlockGravel.java
@@ -18,12 +18,12 @@ public class BlockGravel extends BlockFallable {
     }
 
     @Override
-    public double getHardness() {
-        return 0.6;
+    public float getHardness() {
+        return 0.6f;
     }
 
     @Override
-    public double getResistance() {
+    public float getResistance() {
         return 3;
     }
