+++ src/main/java/cn/nukkit/block/BlockOreGold.java
@@ -17,12 +17,12 @@ public class BlockOreGold extends BlockSolid {
     }
 
     @Override
-    public double getHardness() {
+    public float getHardness() {
         return 3;
     }
 
     @Override
-    public double getResistance() {
+    public float getResistance() {
         return 15;
     }
